v0.1.3
added:
edit option

changes:
visual changes(html, css)
small model,repo and services changes

+++ src/main/java/com/Maruszak/MantisKeeper/model/User.java
@@ -19,7 +19,7 @@ public class User implements UserDetails {
 
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE)
-    private int id;
+    private long id;
 
     @Column
     @NotNull
@@ -48,14 +48,13 @@ public class User implements UserDetails {
     public User() {
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
-
 
     public String getEmail() {
         return email;
